Move template version queries to templates
