--- site/src/api/queries/templateVersions.ts
@@ -1,15 +1,0 @@
-import * as API from "api/api";
-
-export const templateVersionLogs = (versionId: string) => {
-  return {
-    queryKey: ["templateVersion", versionId, "logs"],
-    queryFn: () => API.getTemplateVersionLogs(versionId),
-  };
-};
-
-export const richParameters = (versionId: string) => {
-  return {
-    queryKey: ["templateVersion", versionId, "richParameters"],
-    queryFn: () => API.getTemplateVersionRichParameters(versionId),
-  };
-};
