+++ site/src/pages/CreateWorkspacePage/CreateWorkspacePage.tsx
@@ -22,11 +22,11 @@ import { useMutation, useQuery, useQueryClient } from "react-query";
 import {
   templateByName,
   templateVersionExternalAuth,
+  richParameters,
 } from "api/queries/templates";
 import { autoCreateWorkspace, createWorkspace } from "api/queries/workspaces";
 import { checkAuthorization } from "api/queries/authCheck";
 import { CreateWSPermissions, createWorkspaceChecks } from "./permissions";
-import { richParameters } from "api/queries/templateVersions";
 import { paramsUsedToCreateWorkspace } from "utils/workspace";
 import { useEffectEvent } from "hooks/hookPolyfills";
